bug, fixed queryParams error
