+++ components/stripe/BookingCheckoutPage.tsx
@@ -1,26 +1,22 @@
+"use client";
+
 import React from "react";
+import { useSearchParams } from "next/navigation";
 import { Calendar, Clock, MapPin } from "lucide-react";
 import StarRatingReview from "../ProviderOverallRatingInfo";
 import IconLeftTwoTextRight from "../IconLeftTwoTextRight";
 
-type BookingCheckoutPageType = {
-  providerName: string;
-  date: string;
-  time: string;
-  location: string;
-  service: string;
-  serviceDuration: string;
-  serviceCost: string;
-};
-export default function BookingCheckoutPage({
-  providerName,
-  date,
-  time,
-  location,
-  service,
-  serviceDuration,
-  serviceCost,
-}: BookingCheckoutPageType) {
+export default function BookingCheckoutPage() {
+  const searchParams = useSearchParams();
+  const providerName = searchParams.get("providername") || "Green Thumb Pros";
+  const date = searchParams.get("date") || "Monday, July 15th, 2025";
+  const time = searchParams.get("time") || "11:00 AM";
+  const location =
+    searchParams.get("location") || "123 Main St, San Francisco, CA 94102";
+  const service = searchParams.get("service") || "Lawn Mowing";
+  const serviceDuration = searchParams.get("serviceduration") || "60 minutes";
+  const serviceCost = searchParams.get("servicecost") || "65";
+
   return (
     <section className="mb-10">
       <div className="rounded-lg border-1 border-light-accent bg-white p-4">
